requirements.txt
requests>=2.28.0
numpy>=1.21.0
faiss-cpu>=1.7.0
